fix: PlayerState 金额字段 float64→int64(元) + 冻结RowsAffected检查 + Broker ≥90%现价抛售
- domain: Cash/FrozenCash/MarginDebt/Order.FrozenAmount/AssetLog → int64
- config: StartingCash=100_000, MinCommission=5 (去 .0)
- store: 金额签名全改 int64, FreezeCash/UnfreezeCash/DeductFrozenCash 检查 RowsAffected
- engine: centsToYuan 四舍五入, calcCommission/calcStampTax Ceil取整
- engine/broker: 删重复 tx.Rollback, 买单≥现价90%才释放库存
- handler/ws: 响应/消息字段同步 int64
- docs: 记录资金模型更新

+++ jjs-server/internal/domain/models.go
@@ -20,9 +20,9 @@ type User struct {
 type PlayerState struct {
 	PlayerID   string  `gorm:"type:varchar(12);primaryKey"`
 	Nickname   string  `gorm:"type:varchar(50);not null;default:''"`
-	Cash       float64 `gorm:"not null;default:0"`
-	FrozenCash float64 `gorm:"not null;default:0"`
-	MarginDebt float64 `gorm:"not null;default:0"`
+	Cash       int64 `gorm:"not null;default:0"`
+	FrozenCash int64 `gorm:"not null;default:0"`
+	MarginDebt int64 `gorm:"not null;default:0"`
 }
 
 type Company struct {
@@ -93,8 +93,8 @@ type AssetLog struct {
 	ID        uint      `gorm:"primaryKey;autoIncrement"`
 	PlayerID  string    `gorm:"type:varchar(12);index;not null"`
 	Type      string    `gorm:"type:varchar(20);not null"`
-	Amount    float64   `gorm:"not null"`
-	Balance   float64   `gorm:"not null"`
+	Amount    int64     `gorm:"not null"`
+	Balance   int64     `gorm:"not null"`
 	Note      string    `gorm:"type:varchar(200);default:''"`
 	CreatedAt time.Time `gorm:"autoCreateTime"`
 }
@@ -129,7 +129,7 @@ type Order struct {
 	FilledQty    int64     `gorm:"not null;default:0"                                 json:"filled_qty"`
 	Status       string    `gorm:"type:varchar(20);not null;default:'open'"           json:"status"`
 	SeqNum       int64     `gorm:"not null;default:0"                                 json:"-"`
-	FrozenAmount float64   `gorm:"not null;default:0"                                 json:"-"`
+	FrozenAmount int64     `gorm:"not null;default:0"                                 json:"-"`
 	CreatedAt    time.Time `gorm:"autoCreateTime"                                     json:"created_at"`
 	UpdatedAt    time.Time `gorm:"autoUpdateTime"                                     json:"-"`
 }
